+++ src/main/java/shop/wannab/book_service/book/controller/AdminBookController.java
@@ -1,14 +1,16 @@
 package shop.wannab.book_service.book.controller;
 
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import shop.wannab.book_service.book.dto.request.BookCreateRequest;
-import shop.wannab.book_service.book.dto.request.BookUpdateRequest;
-import shop.wannab.book_service.book.dto.response.BookListResponse;
+import shop.wannab.book_service.book.controller.request.AladinBookCreateRequest;
+import shop.wannab.book_service.book.controller.request.BookCreateRequest;
+import shop.wannab.book_service.book.controller.request.BookUpdateRequest;
+import shop.wannab.book_service.book.controller.response.BookListResponse;
 import shop.wannab.book_service.book.service.AdminBookService;
 import shop.wannab.book_service.global.response.ApiResponse;
 
@@ -19,7 +21,10 @@ public class AdminBookController {
 
     private final AdminBookService adminBookService;
 
-    // 도서 목록 검색(리스트)
+    /**
+     * 도서 목록 리스트
+     * @param pageable 페이징 객체
+     */
     @GetMapping
     public ResponseEntity<ApiResponse<Page<BookListResponse>>> getBooks(
             @PageableDefault(size = 20) Pageable pageable) {
@@ -27,14 +32,21 @@ public class AdminBookController {
         return ResponseEntity.ok(ApiResponse.success(bookList));
     }
 
-    // 도서 등록
+    /**
+     * 도서 등록
+     * @param request 도서 등록 정보
+     */
     @PostMapping
-    public ResponseEntity<ApiResponse<Void>> createBook(@RequestBody BookCreateRequest request){
+    public ResponseEntity<ApiResponse<Void>> createBook(@RequestBody @Valid BookCreateRequest request){
         adminBookService.createBook(request);
         return ResponseEntity.ok(ApiResponse.success(null));
     }
 
-    // 도서 수정
+    /**
+     * 도서 수정
+     * @param bookId 도서 PK
+     * @param request 도서 수정 정보
+     */
     @PutMapping("/{book-id}")
     public ResponseEntity<ApiResponse<Void>> updateBook(@PathVariable("book-id") Long bookId,
                                                         @RequestBody BookUpdateRequest request){
@@ -42,10 +54,23 @@ public class AdminBookController {
         return ResponseEntity.ok(ApiResponse.success(null));
     }
 
-    // 도서 삭제
+    /**
+     * 도서 삭제
+     * @param bookId 도서 PK
+     */
     @DeleteMapping("/{book-id}")
     public ResponseEntity<ApiResponse<Void>> deleteBook(@PathVariable("book-id") Long bookId){
         adminBookService.deleteBook(bookId);
+        return ResponseEntity.ok(ApiResponse.success(null));
+    }
+
+    /**
+     * 알라디 도서 등록
+     * @param request 알라딘 도서 등록 객체
+     */
+    @PostMapping("/aladin")
+    public ResponseEntity<ApiResponse<Void>> createAladinBook(@RequestBody AladinBookCreateRequest request){
+        adminBookService.aladinCreateBook(request);
         return ResponseEntity.ok(ApiResponse.success(null));
     }
 }
