+++ src/main/java/shop/wannab/book_service/book/controller/request/BookCreateRequest.java
@@ -1,11 +1,13 @@
-package shop.wannab.book_service.book.dto.request;
+package shop.wannab.book_service.book.controller.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import shop.wannab.book_service.book.entity.Book;
 import shop.wannab.book_service.global.deserializer.CommaSeparatedToListDeserializer;
 
 import java.time.LocalDate;
@@ -16,14 +18,14 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class BookUpdateRequest {
+public class BookCreateRequest {
     private String title;
     private String description;
     private LocalDate publicationDate;
     private Integer originPrice;
     private Integer stock;
     private String bookChapter;
-    private String isbn;
+    @NotBlank private String isbn;
     private boolean status;
 
     @JsonDeserialize(using = CommaSeparatedToListDeserializer.class)
@@ -37,4 +39,18 @@ public class BookUpdateRequest {
 
     @JsonDeserialize(using = CommaSeparatedToListDeserializer.class)
     private List<String> bookImages;
+
+
+    public Book toEntity(){
+        return Book.builder()
+                .title(this.title)
+                .description(this.description)
+                .publicationDate(this.publicationDate)
+                .isbn(this.isbn)
+                .originPrice(this.originPrice)
+                .stock(this.stock)
+                .status(this.status)
+                .bookChapter(this.bookChapter)
+                .build();
+    }
 }
