+++ src/main/java/shop/wannab/book_service/review/controller/request/ReviewCreateRequest.java
@@ -1,4 +1,4 @@
-package shop.wannab.book_service.review.dto.request;
+package shop.wannab.book_service.review.controller.request;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
@@ -6,18 +6,20 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import shop.wannab.book_service.global.deserializer.CommaSeparatedToListDeserializer;
+import shop.wannab.book_service.review.entity.Review;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Builder
 @Getter
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReviewUpdateRequest {
+public class ReviewCreateRequest {
     private String reviewContent;
     private Integer reviewScore;
-    private LocalDateTime reviewUpdatedAt;
+    private LocalDateTime reviewCreatedAt;
+    private Long obId;
 
     @JsonDeserialize(using = CommaSeparatedToListDeserializer.class)
     private List<String> reviewImages;
