+++ src/main/java/shop/wannab/book_service/review/controller/response/BookReviewListResponse.java
@@ -1,4 +1,4 @@
-package shop.wannab.book_service.review.dto.response;
+package shop.wannab.book_service.review.controller.response;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -7,24 +7,25 @@ import shop.wannab.book_service.review.entity.Review;
 import java.time.LocalDateTime;
 import java.util.List;
 
-@Getter
 @Builder
-public class UserReviewListResponse {
+@Getter
+public class BookReviewListResponse {
     private String reviewContent;
     private Integer reviewScore;
     private LocalDateTime reviewCreatedAt;
+    private String username;
     private List<ReviewImageResponse> reviewImages;
 
-    public static UserReviewListResponse from(Review review){
-        return UserReviewListResponse.builder()
+    public static BookReviewListResponse from(Review review, String username){
+        return BookReviewListResponse.builder()
                 .reviewContent(review.getReviewContent())
                 .reviewScore(review.getReviewScore())
                 .reviewCreatedAt(review.getReviewCreatedAt())
+                .username(username)
                 .reviewImages(
                         review.getReviewImages().stream()
                                 .map(ReviewImageResponse::from)
-                                .toList()
-                )
+                                .toList())
                 .build();
     }
 }
