+++ src/main/java/shop/wannab/book_service/review/service/ReviewService.java
@@ -3,10 +3,10 @@ package shop.wannab.book_service.review.service;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import shop.wannab.book_service.review.dto.request.ReviewCreateRequest;
-import shop.wannab.book_service.review.dto.request.ReviewUpdateRequest;
-import shop.wannab.book_service.review.dto.response.BookReviewListResponse;
-import shop.wannab.book_service.review.dto.response.UserReviewListResponse;
+import shop.wannab.book_service.review.controller.request.ReviewCreateRequest;
+import shop.wannab.book_service.review.controller.request.ReviewUpdateRequest;
+import shop.wannab.book_service.review.controller.response.BookReviewListResponse;
+import shop.wannab.book_service.review.controller.response.UserReviewListResponse;
 
 public interface ReviewService {
     Page<BookReviewListResponse> getBookReviewList(Pageable pageable, Long bookId);
