+++ src/main/java/shop/wannab/book_service/review/service/impl/ReviewServiceImpl.java
@@ -11,10 +11,10 @@ import shop.wannab.book_service.book.exception.BookErrorCode;
 import shop.wannab.book_service.book.repository.BookRepository;
 import shop.wannab.book_service.client.UserClient;
 import shop.wannab.book_service.client.dto.response.UserResponseWrapper;
-import shop.wannab.book_service.review.dto.request.ReviewCreateRequest;
-import shop.wannab.book_service.review.dto.request.ReviewUpdateRequest;
-import shop.wannab.book_service.review.dto.response.BookReviewListResponse;
-import shop.wannab.book_service.review.dto.response.UserReviewListResponse;
+import shop.wannab.book_service.review.controller.request.ReviewCreateRequest;
+import shop.wannab.book_service.review.controller.request.ReviewUpdateRequest;
+import shop.wannab.book_service.review.controller.response.BookReviewListResponse;
+import shop.wannab.book_service.review.controller.response.UserReviewListResponse;
 import shop.wannab.book_service.review.entity.Review;
 import shop.wannab.book_service.review.entity.ReviewImage;
 import shop.wannab.book_service.review.exception.ReviewApiException;
